+++ app/src/main/java/org/oxycblt/auxio/music/MusicSettings.kt
@@ -44,8 +44,9 @@ interface MusicSettings : Settings<MusicSettings.Listener> {
     val shouldBeObserving: Boolean
     /** A [String] of characters representing the desired characters to denote multi-value tags. */
     var multiValueSeparators: String
-    /** Whether to trim english articles with song sort names. */
-    val automaticSortNames: Boolean
+    /** Whether to enable more advanced sorting by articles and numbers. */
+    val intelligentSorting: Boolean
+    // TODO: Move sort settings to list module
     /** The [Sort] mode used in [Song] lists. */
     var songSort: Sort
     /** The [Sort] mode used in [Album] lists. */
@@ -110,7 +111,7 @@ class MusicSettingsImpl @Inject constructor(@ApplicationContext context: Context
             }
         }
 
-    override val automaticSortNames: Boolean
+    override val intelligentSorting: Boolean
         get() = sharedPreferences.getBoolean(getString(R.string.set_key_auto_sort_names), true)
 
     override var songSort: Sort