+++ app/src/test/java/org/oxycblt/auxio/music/FakeMusicSettings.kt
@@ -34,7 +34,7 @@ open class FakeMusicSettings : MusicSettings {
     override var multiValueSeparators: String
         get() = throw NotImplementedError()
         set(_) = throw NotImplementedError()
-    override val automaticSortNames: Boolean
+    override val intelligentSorting: Boolean
         get() = throw NotImplementedError()
     override var songSort: Sort
         get() = throw NotImplementedError()