+++ CHANGELOG.md
@@ -21,6 +21,9 @@ rather than on start/stop
 deletion
 - Fix "format" not appearing in song properties view
 
+#### What's Changed
+- "Ignore articles when sorting" is now "Intelligent sorting"
+
 ## 3.0.3
 
 #### What's New